Major changes done, exp render bug fixed Most of the functionalities are implemented, some still clean up and improvements, but questions are being sent and received, as well as answers being saved to the correct players.
Bug fixes and lessons:
➡  The exponentially re-renders are a symptom of React saving a history of
state changes that are not cleared at each render, hence exponentially
growing the number of renders at each render. In React without hooks a
componentDidUnmount or componentDidUpdate function was used to clear and
control this history, in hooks an array is passed as second argument of
useEffect triggering the function only when a state variable changes,
thus fixing the bug in question.

➡  In the return function that established what is being rendered, some
components receive props, and those props can be functions, while it
would be intuitive to pass as prop a function with argument as for
example:

<Component trigger={aFunction(arg1, arg2)}>

this will actually execute the function at render, and not at trigger
command, the correct way is to pass the function as a variable and find
another way of passing the arguments to the function, so correctly
should be just:

<Component trigger={aFunction}>